contrib/cloud.google.com/go: add shared tracing package
