+++ contrib/cloud.google.com/go/pubsub.v1/pubsub.go
@@ -9,11 +9,19 @@ package pubsub
 import (
 	"context"
 
-	"github.com/DataDog/dd-trace-go/contrib/cloud.google.com/go/pubsub.v1/v2/internal/tracing"
-	_ "github.com/DataDog/dd-trace-go/v2/instrumentation" // Blank import to pass TestIntegrationEnabled test
+	"github.com/DataDog/dd-trace-go/v2/contrib/cloud.google.com/go/pubsubtrace"
+	"github.com/DataDog/dd-trace-go/v2/instrumentation"
 
 	"cloud.google.com/go/pubsub"
 )
+
+const componentName = instrumentation.PackageGCPPubsub
+
+var instr *instrumentation.Instrumentation
+
+func init() {
+	instr = instrumentation.Load(componentName)
+}
 
 // Publish publishes a message on the specified topic and returns a PublishResult.
 // This function is functionally equivalent to t.Publish(ctx, msg), but it also starts a publish
@@ -23,7 +31,7 @@ import (
 // the span.
 func Publish(ctx context.Context, t *pubsub.Topic, msg *pubsub.Message, opts ...Option) *PublishResult {
 	traceMsg := newTraceMessage(msg)
-	ctx, closeSpan := tracing.TracePublish(ctx, t, traceMsg, opts...)
+	ctx, closeSpan := pubsubtrace.TracePublish(ctx, t, traceMsg, instr, componentName, opts...)
 	msg.Attributes = traceMsg.Attributes
 
 	return &PublishResult{
@@ -50,7 +58,7 @@ func (r *PublishResult) Get(ctx context.Context) (string, error) {
 // extracts any tracing metadata attached to the received message, and starts a
 // receive span.
 func WrapReceiveHandler(s *pubsub.Subscription, f func(context.Context, *pubsub.Message), opts ...Option) func(context.Context, *pubsub.Message) {
-	traceFn := tracing.TraceReceiveFunc(s, opts...)
+	traceFn := pubsubtrace.TraceReceiveFunc(s, instr, componentName, opts...)
 	return func(ctx context.Context, msg *pubsub.Message) {
 		ctx, closeSpan := traceFn(ctx, newTraceMessage(msg))
 		defer closeSpan()
@@ -58,11 +66,11 @@ func WrapReceiveHandler(s *pubsub.Subscription, f func(context.Context, *pubsub.
 	}
 }
 
-func newTraceMessage(msg *pubsub.Message) *tracing.Message {
+func newTraceMessage(msg *pubsub.Message) *pubsubtrace.Message {
 	if msg == nil {
 		return nil
 	}
-	return &tracing.Message{
+	return &pubsubtrace.Message{
 		ID:              msg.ID,
 		Data:            msg.Data,
 		OrderingKey:     msg.OrderingKey,
