+++ contrib/cloud.google.com/go/pubsubtrace/tracing.go
@@ -1,16 +1,16 @@
 // Unless explicitly stated otherwise all files in this repository are licensed
 // under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2024 Datadog, Inc.
+// Copyright 2025 Datadog, Inc.
 
-// Package tracing contains tracing logic for the cloud.google.com/go/pubsub.v1 instrumentation.
+// Package tracing contains tracing logic for the cloud.google.com/go/pubsub instrumentation.
 //
-// WARNING: this package SHOULD NOT import cloud.google.com/go/pubsub.
+// WARNING: this package SHOULD NOT import cloud.google.com/go/pubsub or cloud.google.com/go/pubsub/v2.
 //
 // The motivation of this package is to support orchestrion, which cannot use the main package because it imports
-// the cloud.google.com/go/pubsub package, and since orchestrion modifies the library code itself,
+// the package, and since orchestrion modifies the library code itself,
 // this would cause an import cycle.
-package tracing
+package pubsubtrace
 
 import (
 	"context"
@@ -21,8 +21,6 @@ import (
 	"github.com/DataDog/dd-trace-go/v2/ddtrace/tracer"
 	"github.com/DataDog/dd-trace-go/v2/instrumentation"
 )
-
-const componentName = instrumentation.PackageGCPPubsub
 
 type Message struct {
 	ID              string
@@ -41,8 +39,8 @@ type Subscription interface {
 	String() string
 }
 
-func TracePublish(ctx context.Context, topic Topic, msg *Message, opts ...Option) (context.Context, func(serverID string, err error)) {
-	cfg := defaultConfig()
+func TracePublish(ctx context.Context, topic Topic, msg *Message, instr *instrumentation.Instrumentation, componentName instrumentation.Package, opts ...Option) (context.Context, func(serverID string, err error)) {
+	cfg := defaultConfig(instr)
 	for _, opt := range opts {
 		opt.apply(cfg)
 	}
@@ -70,7 +68,7 @@ func TracePublish(ctx context.Context, topic Topic, msg *Message, opts ...Option
 		msg.Attributes = make(map[string]string)
 	}
 	if err := tracer.Inject(span.Context(), tracer.TextMapCarrier(msg.Attributes)); err != nil {
-		instr.Logger().Debug("contrib/cloud.google.com/go/pubsub.v1/trace: failed injecting tracing attributes: %s", err.Error())
+		instr.Logger().Debug("contrib/cloud.google.com/go/pubsubtrace: failed injecting tracing attributes: %s", err.Error())
 	}
 	span.SetTag("num_attributes", len(msg.Attributes))
 
@@ -84,12 +82,12 @@ func TracePublish(ctx context.Context, topic Topic, msg *Message, opts ...Option
 	return ctx, closeSpan
 }
 
-func TraceReceiveFunc(s Subscription, opts ...Option) func(ctx context.Context, msg *Message) (context.Context, func()) {
-	cfg := defaultConfig()
+func TraceReceiveFunc(s Subscription, instr *instrumentation.Instrumentation, componentName instrumentation.Package, opts ...Option) func(ctx context.Context, msg *Message) (context.Context, func()) {
+	cfg := defaultConfig(instr)
 	for _, opt := range opts {
 		opt.apply(cfg)
 	}
-	instr.Logger().Debug("contrib/cloud.google.com/go/pubsub.v1/trace: Wrapping Receive Handler: %#v", cfg)
+	instr.Logger().Debug("contrib/cloud.google.com/go/pubsubtrace: Wrapping Receive Handler: %#v", cfg)
 	return func(ctx context.Context, msg *Message) (context.Context, func()) {
 		parentSpanCtx, _ := tracer.Extract(tracer.TextMapCarrier(msg.Attributes))
 		opts := []tracer.StartSpanOption{
